Add files via upload
The objectives of the analysis were met as the model techniques used (ridge and random forests) predicted the house prices in India. Random forests method generated better house price predictions than ridge regression because it has additional and sophisticate features which deeply analyze the data. Input variables such as the grade of the house, living area, area of the house, latitudes, and longitudes play a major role in determining the house prices. A validation score of 87.27% was obtain from random forests model. This implies that the model is good and can be used to make decisions and policies on housing. However, the accuracy score of house prices prediction can be further improved by applying other techniques like XGBoost gradient descent, and hyperparameter tunings as used by Osman et al., (2021) to predict the groundwater levels in Selangor Malaysia.